Make the directory containing aspect data a pipeline context parameter.
These files are too big to include in git, and therefore also too big
to include in a wheel, and so we shouldn’t be expecting them to be
embedded in the installation.

To do this cleanly, aspect_tables() is generalized so it can take
general filters and any other arguments to parquet.read_table that may
be necessary.  Also, aspect.py is renamed to aspect/__init__.py so
that wheels will contain the aspect directory, in case someone wants
to install the wheel somewhere and then copy the aspect data into that
directory instead of passing down some other location.

No existing usage should be affected.