Include the entire contents of the package in built sdists and wheels.
Apparently `include = ["gPhoton"]` in the `setuptools.packages.find`
section of `pyproject.toml` means to include the top-level “gPhoton”
package in sdists and wheels but *not* any sub-packages, which in
this case meant a whole lot of important code was getting left out.

Also, for calibration files to be included in sdists and wheels,
a `setuptools.package-data` section is required.