Fix all existing type-checking errors from mypy.
This does *not* make gPhoton fully typed; in fact, in a couple of
places I *removed* type annotations because that was the path of least
resistance.  `mypy --strict` reports 576 errors, most of which are
missing type annotations.

However, mypy *without* --strict now reports *zero* errors, which is
worthwhile just because I will now be able to make mypy tell me if I
forgot to add an `aspect_dir` argument to anything, without being
buried under all the preexisting issues.

This *does* flush out a grand total of two real bugs (both involving
a high-level function that advertised itself as accepting a Pathlike
passing that Pathlike down to a low-level function that only accepted
a str) and also, IMO, makes the code easier to understand in several
places (due to less convoluted control flow).