Fix indentation error in gPhoton.cals, make byteswap platform byteorder agnostic
FITS numeric types are always MSB. Some libraries we use (especially numba)
behave unreliably with ndarrays of non-native byteorder, so we formerly byte-swapped all
loaded FITS arrays. A typo in a previous commit skipped this byteswap for
non-structured ndarrays. This commit replaces the byteswapping behavior, and
attempts to ensure that we do not swap to the _wrong_ byteorder if someone
happens to run gPhoton2 on a big-endian system.