Reorganize generation of data file paths from scratch.
gPhoton.eclipse now exposes a set of functions that generate
individual paths, taking only the specific arguments that are needed
for that path.  These functions *do not* accept a ‘root’ argument;
they always produce a relative path, which the caller should combine
with whatever it wants the data root to be.

eclipse_to_paths remains, but now it returns a custom read-only
mapping class that lazily calculates only the paths you look up,
because most of the time only one path is used.  Its interface
*should be* unchanged (including the ‘root’ argument).

The functionality of emojified() has been folded into the individual
path generation functions.