Enable ‘W’ lint group.  Add pre-commit configuration.
The ‘W’ lint group checks for invisible, mechanically correctible
whitespace errors: trailing horizontal whitespace, wrong number of
‘\n’ characters at the end of a file, and mixed spaces and tabs.

Automate correction of these issues with an opt-in pre-commit hook.
If you turn it on, it will automatically fix these for you when you
do ‘git commit’.  It will also automatically run all the lints, but
it won’t fix those for you (even if it can).  To turn it on see the
instructions at <https://pre-commit.com/#install>.

The .tbl files in cal_data are excluded from correction; they appear
to have *intentional* trailing whitespace.