Enable ‘RET’ lint group.
The thing we actually *want* from this lint group is its detection of
functions that sometimes explicitly return a value, and other times,
implicitly return None (either by falling off the end, or by using a
‘return’ statement with no argument).  This is suspicious because
these functions’ callers might not be expecting them to return None.
In this commit we just change them all to explicitly return None,
so the overall behavior does not change.  Down the road we will
hopefully be able to catch the caller/callee mismatch using type
annotations.

The lint group also complains about the construct

    def loose(...):
        ...
        val = compute_return_value(...)
        return val

in which ‘val’ can be eliminated:

    def tight(...):
        ...
        return compute_return_value(...)

This is not something I care about one way or the other, but the
existing code *usually* does the ‘tight’ thing, so I’m fine with
enforcing consistency.