The ‘DTZ’ lint group turns out to be undesirable.
This lint group tries to get people to not use “naive” datetime
objects; instead they should always consistently supply a time zone to
the constructors.  The problem with this idea is that the stdlib
doesn’t have any way to express “whatever the user’s time zone is”
other than by using naive datetime objects.

The only thing we do with datetime objects is generate log messages,
so we *want* “whatever the user’s time zone is,” and there’s no real
value in getting that from the third-party modules that deal with
timezones more cleverly than the stdlib does.  Naive datetime objects
are the Right Thing for us.

This lint group did bring to my attention an inconsistency in
pretty.py, where sometimes timestamps were being forced to UTC,
so that’s fixed in this patch.