Enable most of ‘SIM’ lint group.
The lints in this group that matter to us are use of ‘open’ without a
context manager (which risks leaking the open file) and unnecessarily
nested if statements.

Switching from ‘dict.keys()’ to bare ‘dict’ is debatable, but it found
a bunch of places where we ought to have been using ‘dict.items()’ or
‘dict.values()’ instead so I’m calling that a win.

Don’t enable SIM108, which very aggressively recommends collapsing

    if condition:
        var = value1
    else:
        var = value2

to

    var = value1 if condition else value2

I feel that this harms readability when the expressions involved
aren’t very simple, and in gPhoton they are usually complicated.