Enable ‘BLE’ lint group.
This is just one lint, complaining about catch-all exception clauses
that throw away the exception.  It turns out that there’s only one
place in gPhoton where we do that and it’s on purpose, but it’d be
an easy thing to introduce *not* on purpose, so enable it and tag
the on-purpose place as such.