Move ‘FBT’ lint group to the “TODO: with typing” category.
This lint group wants you to not have positional, optional, boolean
arguments, which is just barely on the desirable side of the line
between lints that catch bugs and lints that just enforce a coding
style.  The changes required to comply with it are basically to
convert a whole bunch of positional optional arguments to
keyword-only, and this is too risky until we can have mypy tell us
where we forgot to fix up callsites.

Also adjust the annotations in ruff.toml to make clearer which lint
groups should be enabled along with strict typing.