Define type variables in types.py for standard ways of using NumPy.
Our low-level code generally expects a bunch of NumPy array and
scalar arguments that all have the same dtype, but it doesn’t
particularly care *which dtype* that is.  Introduce type variables
to make annotating for this usage easier.

I’m not sure the way I’m writing these is 100% accurate but it seems
to work (as used in subsequent commits).

Note that it is not currently possible to specify the expected *shape*
of an ndarray; there is a plan to make this possible (see PEP 646) but
it is not yet implemented in NumPy itself.