Add a wrapper for numba.njit that preserves type annotations.
Numba is currently untyped, and that means that a function decorated
with @numba.(n)jit *loses* its type annotations, because the decorator
could have returned *anything*.  Introduce a wrapper, numba_utilz.jit,
that informs type checkers that the decorator applies to a callable
and returns a callable with the same signature.  This is enough to
make type checkers see through the decoration.

The wrapper is called ‘jit’ rather than ‘njit’ because numba.njit is
a mostly undocumented wrapper around numba.jit that just forces the
keyword arguments `nopython=True` and `forceobj=False`.  Our wrapper
can do the same.  It also sets `cache=True` because we always do that.

The wrapper does not attempt to be any more specific about the type
signature of the decorator itself, than what is necessary to make
this work.  That may not even be *possible* while Numba is untyped.