Add full type annotations to io/fits_utils.py; partially clean up.
This file is a shim over several untyped libraries, so there’s a
lot of casting going on, alas.

This file is still in need of refactoring, but the way AgnosticHDU and
AgnosticHDUL work is now quite a bit tighter; in particular,
crudely_find_library is no longer necessary at all and has been
removed.

AgnosticHDUL can now be used in ‘with’ blocks.  In two places,
direct use of close() has been replaced with a ‘with’ block.