Add full type annotations to coadd.py.
This includes a limited number of local cleanups, notably to
bounds_from_corners, flag_mask, _check_oob, and the cut_skyboxes
internal calling convention.  Further cleanups are possible once
we’re in a position to start introducing data classes.

Typing this file revealed an error in the typing of
coords.wcs.sky_box_to_image_box: too picky on the input side and too
generic on the output side.