Add full type annotations to gPhoton/moviemaker/_steps.py.
This code uses dicts as pseudo-records a lot and therefore the typing
is sloppier than I would ideally like.

For clarity, sm_make_map was folded into sm_make_maps and the loop
unrolled.

So as not to be using one dict for two radically different things, up
in make_movies, the alternative threads/no-threads code paths needed
to be split apart.

Some of the functions in calibrate.py were incorrectly annotated in
an earlier commit; this is corrected now I can see what their callers
actually supply.

In a few places we use the experimental shape typing from numpy 2.2;
this should be removed with prejudice if it causes any problems
whatsoever (see https://github.com/numpy/numpy/issues/28076 ) but
it does seem to work for the very limited case this code wants, i.e.
“this is a 2-d matrix”.